Create A vs B Testing for ShoeFly.com
ShoeFly.com have two different versions of an ad, which they have placed in emails, as well as in banner ads on Facebook, Twitter, and Google. ShoeFly.com want to know how the two ads are performing on each of the different platforms on each day of the week.  I help them analyze the data using aggregate measures.